[S08P31A405-387] Fe/Fix: fix errors
앱 아이콘/런처 추가, 배경화면 색상 변경, 로딩창 에러 수정


Former-commit-id: 8c9e9f918ca11a781c3a1bff83a18817d3181250

+++ frontend/mmARt/app/src/main/java/com/example/mmart/Retrofit.kt
@@ -211,7 +211,7 @@ data class IdxReviewsResult(
 
 data class ReviewsPos(
     val reviewRes: List<ReviewDetail>,
-    val pos: Int
+    val pos: Float
 )
 
 // 유저 별 리뷰 조회
